Use plugin-content-docs instead of duplicating old code + fixes and features
This fixes a lot of issues like auto generated sidebars and support for all new sidebar features like doc links on categories, as well as sidebar info specified in the md files header.
It also simplifies version handling.

Separated document titles from document ids as the id might not always be the same as the title. And it also fixes the download button not finding the files if the title/id/url doesn't match as is possible with the slug option on sidebar entries.
Fixed toc page numbering for unnumbered headings.

Added options for:
- useExtraPaths to use with the express server.
- alwaysIncludeSiteStyles to be able to add both custom styles and the docusaurus ones.
- ignoreCssSelectors to be able to remove html elements before generating the pdf.
- margins and coverMargins as they we're hardcoded.

Removed rootDocIds as it can be deduced from the docs plugin data.

A drawback to this new approach is that we depend on data passed to postBuild, so we can't support the custom command anymore. I added a haxx env var to trigger the build if autoBuildPdfs is false.